update grad handler
